Started Parser.expr, added multiple Token methods, and fixed two bugs
- Parser's "expr" method can transform an infix expression into a
post-fix expression. It can't yet evaluate the post-fix expression
- Added Token's "toResultValue" method which returns a ResultValue with
the values from the current token.
- Added Token's "getPrecedence" method which will return the correct
precedence value for both token and stack precedences. This method is
only called by "precedence" and "stackPrecdence" methods which pass a
boolean indicating which precedence to return (I consolidated the
precedences to one method to improve both readability and debugging)
- Fixed a bug where a comment on the last line would cause an
IndexOutOfBoundsException...
- Fixed a bug where a token for a variable which was previously declared
was not being initialized with the correct classifications