Re-factored Utility and added calls to those methods from the parser
- Parser now calls most of the Utility methods.
- Utility methods now only take ResultValue objects and create Numerics
if needed.
- Utility methods currently coerce the second operand to the type of the
first if needed, and they error out if the type of the first operand is
not valid for the operation. This may change to coercing the first type
whenever possible (need to meet with Clark to discuss). Because of this,
comments for those methods haven't been finished yet.
- Edited the 'coerce' method in Utility to coerce a ResultValue to the
given type if possible or provide an appropriate error message if not
possible.